+++ source/packages/@aws-accelerator/constructs/lib/aws-guardduty/create-members/index.ts
@@ -25,7 +25,7 @@ import {
   OrganizationFeatureConfiguration,
   BadRequestException,
 } from '@aws-sdk/client-guardduty';
-import { OrganizationsClient, ListAccountsCommand } from '@aws-sdk/client-organizations';
+import { OrganizationsClient, ListAccountsCommand, ListAccountsCommandOutput } from '@aws-sdk/client-organizations';
 import { CloudFormationCustomResourceEvent } from '@aws-accelerator/utils/lib/common-types';
 
 /**
@@ -43,8 +43,11 @@ export async function handler(event: CloudFormationCustomResourceEvent): Promise
 > {
   const region = event.ResourceProperties['region'];
   const partition = event.ResourceProperties['partition'];
+  const guardDutyMemberAccountIds: string[] = event.ResourceProperties['guardDutyMemberAccountIds'];
   const enableS3Protection: boolean = event.ResourceProperties['enableS3Protection'] === 'true';
   const enableEksProtection: boolean = event.ResourceProperties['enableEksProtection'] === 'true';
+  const autoEnableOrgMembersFlag: boolean = event.ResourceProperties['autoEnableOrgMembers'] === 'true';
+
   const solutionId = process.env['SOLUTION_ID'];
   const chunkSize = process.env['CHUNK_SIZE'] ? parseInt(process.env['CHUNK_SIZE']) : 50;
 
@@ -63,12 +66,33 @@ export async function handler(event: CloudFormationCustomResourceEvent): Promise
 
   let nextToken: string | undefined = undefined;
 
+  let existingMemberAccountIds: string[] = [];
+
+  let autoEnableOrgMembers: 'ALL' | 'NEW' | 'NONE' = 'ALL';
+  if (!autoEnableOrgMembersFlag) {
+    autoEnableOrgMembers = 'NONE';
+  }
+
   console.log(`EnableS3Protection: ${enableS3Protection}`);
   console.log(`EnableEksProtection: ${enableEksProtection}`);
+  console.log(`autoEnableOrgMembers: ${autoEnableOrgMembers}`);
 
   switch (event.RequestType) {
     case 'Create':
     case 'Update':
+      const features = getOrganizationFeaturesEnabled(enableS3Protection, enableEksProtection);
+
+      console.log('starting - UpdateOrganizationConfiguration');
+      try {
+        autoEnableOrgMembersFlag
+          ? await updateOrganizationConfiguration(guardDutyClient, detectorId!, autoEnableOrgMembers, features)
+          : await updateOrganizationConfiguration(guardDutyClient, detectorId!, autoEnableOrgMembers);
+
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      } catch (e: any) {
+        return { Status: 'Failure', StatusCode: e.statusCode };
+      }
+
       console.log('starting - CreateMembersCommand');
       const allAccounts: AccountDetail[] = [];
 
@@ -76,9 +100,8 @@ export async function handler(event: CloudFormationCustomResourceEvent): Promise
         const page = await throttlingBackOff(() =>
           organizationsClient.send(new ListAccountsCommand({ NextToken: nextToken })),
         );
-        for (const account of page.Accounts ?? []) {
-          allAccounts.push({ AccountId: account.Id!, Email: account.Email! });
-        }
+
+        allAccounts.push(...getMembersToCreate(page, guardDutyMemberAccountIds));
         nextToken = page.NextToken;
       } while (nextToken);
 
@@ -91,14 +114,21 @@ export async function handler(event: CloudFormationCustomResourceEvent): Promise
         );
       }
 
-      const features = getOrganizationFeaturesEnabled(enableS3Protection, enableEksProtection);
+      // Cleanup members removed from deploymentTarget
+      if (guardDutyMemberAccountIds.length > 0) {
+        console.log('Inititating cleanup of members removed from deploymentTargets');
+        existingMemberAccountIds = await getExistingMembers(guardDutyClient, detectorId!);
 
-      console.log('starting - UpdateOrganizationConfiguration');
-      try {
-        await updateOrganizationConfiguration(guardDutyClient, detectorId!, features, 'ALL');
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      } catch (e: any) {
-        return { Status: 'Failure', StatusCode: e.statusCode };
+        const memberAccountIdsToDelete: string[] = [];
+        for (const accountId of existingMemberAccountIds) {
+          if (!guardDutyMemberAccountIds.includes(accountId)) {
+            memberAccountIdsToDelete.push(accountId);
+          }
+        }
+        if (memberAccountIdsToDelete.length > 0) {
+          const chunkedAccountsForDelete = chunkArray(memberAccountIdsToDelete, chunkSize);
+          await disassociateAndDeleteMembers(guardDutyClient, detectorId!, chunkedAccountsForDelete);
+        }
       }
 
       console.log('Returning Success');
@@ -107,45 +137,89 @@ export async function handler(event: CloudFormationCustomResourceEvent): Promise
     case 'Delete':
       const disabledFeatures = getOrganizationFeaturesEnabled(false, false);
       try {
-        await updateOrganizationConfiguration(guardDutyClient, detectorId!, disabledFeatures, 'NONE');
+        await updateOrganizationConfiguration(guardDutyClient, detectorId!, 'NONE', disabledFeatures);
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } catch (e: any) {
         return { Status: 'Failure', StatusCode: e.statusCode };
       }
 
-      const existingMemberAccountIds: string[] = [];
-      nextToken = undefined;
-      do {
-        const page = await throttlingBackOff(() =>
-          guardDutyClient.send(new ListMembersCommand({ DetectorId: detectorId!, NextToken: nextToken })),
-        );
-        for (const member of page.Members ?? []) {
-          existingMemberAccountIds.push(member.AccountId!);
-        }
-        nextToken = page.NextToken;
-      } while (nextToken);
+      existingMemberAccountIds = await getExistingMembers(guardDutyClient, detectorId!);
 
       if (existingMemberAccountIds.length > 0) {
         const chunkedAccountsForDelete = chunkArray(existingMemberAccountIds, chunkSize);
-
-        for (const existingMemberAccountIdBatch of chunkedAccountsForDelete) {
-          console.log(`Initiating disassociateMembers request for ${existingMemberAccountIdBatch.length} accounts`);
-          await throttlingBackOff(() =>
-            guardDutyClient.send(
-              new DisassociateMembersCommand({ AccountIds: existingMemberAccountIdBatch, DetectorId: detectorId! }),
-            ),
-          );
-
-          console.log(`Initiating deleteMembers request for ${existingMemberAccountIdBatch.length} accounts`);
-          await throttlingBackOff(() =>
-            guardDutyClient.send(
-              new DeleteMembersCommand({ AccountIds: existingMemberAccountIdBatch, DetectorId: detectorId! }),
-            ),
-          );
-        }
+        await disassociateAndDeleteMembers(guardDutyClient, detectorId!, chunkedAccountsForDelete);
       }
 
       return { Status: 'Success', StatusCode: 200 };
+  }
+}
+
+/**
+ * Get accounts details for the GuardDuty members to create
+ * @param page
+ * @param guardDutyMemberAccountIds
+ * @returns
+ */
+function getMembersToCreate(page: ListAccountsCommandOutput, guardDutyMemberAccountIds: string[]): AccountDetail[] {
+  const allAccounts: AccountDetail[] = [];
+  for (const account of page.Accounts ?? []) {
+    if (guardDutyMemberAccountIds.length > 0) {
+      if (guardDutyMemberAccountIds.includes(account.Id!)) {
+        allAccounts.push({ AccountId: account.Id!, Email: account.Email! });
+      }
+    } else {
+      allAccounts.push({ AccountId: account.Id!, Email: account.Email! });
+    }
+  }
+  return allAccounts;
+}
+
+/**
+ * Function to get existing guardduty members
+ * @param guardDutyClient
+ * @param detectorId
+ * @returns string[]
+ */
+async function getExistingMembers(guardDutyClient: GuardDutyClient, detectorId: string): Promise<string[]> {
+  const existingMemberAccountIds: string[] = [];
+  let nextToken: string | undefined = undefined;
+  do {
+    const page = await throttlingBackOff(() =>
+      guardDutyClient.send(new ListMembersCommand({ DetectorId: detectorId, NextToken: nextToken })),
+    );
+    for (const member of page.Members ?? []) {
+      existingMemberAccountIds.push(member.AccountId!);
+    }
+    nextToken = page.NextToken;
+  } while (nextToken);
+
+  return existingMemberAccountIds;
+}
+
+/**
+ * Function to disassociate and delete guardduty members
+ * @param guardDutyClient
+ * @param chunkedAccountsForDelete
+ */
+async function disassociateAndDeleteMembers(
+  guardDutyClient: GuardDutyClient,
+  detectorId: string,
+  chunkedAccountsForDelete: string[][],
+) {
+  for (const existingMemberAccountIdBatch of chunkedAccountsForDelete) {
+    console.log(`Initiating disassociateMembers request for ${existingMemberAccountIdBatch.length} accounts`);
+    await throttlingBackOff(() =>
+      guardDutyClient.send(
+        new DisassociateMembersCommand({ AccountIds: existingMemberAccountIdBatch, DetectorId: detectorId }),
+      ),
+    );
+
+    console.log(`Initiating deleteMembers request for ${existingMemberAccountIdBatch.length} accounts`);
+    await throttlingBackOff(() =>
+      guardDutyClient.send(
+        new DeleteMembersCommand({ AccountIds: existingMemberAccountIdBatch, DetectorId: detectorId }),
+      ),
+    );
   }
 }
 
@@ -181,8 +255,8 @@ async function getDetectorId(guardDutyClient: GuardDutyClient): Promise<string |
 async function updateOrganizationConfiguration(
   guardDutyClient: GuardDutyClient,
   detectorId: string,
-  featureList: OrganizationFeatureConfiguration[],
   autoEnableOrgMembers: 'ALL' | 'NEW' | 'NONE',
+  featureList?: OrganizationFeatureConfiguration[],
 ) {
   try {
     await guardDutyClient.send(
@@ -194,7 +268,7 @@ async function updateOrganizationConfiguration(
     );
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
   } catch (e: any) {
-    if (e instanceof BadRequestException) {
+    if (e instanceof BadRequestException && featureList) {
       console.log('Retrying with only S3 protection');
       const featureListS3Only = featureList.filter(feat => feat.Name === 'S3_DATA_EVENTS');
       await guardDutyClient.send(