+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore(logging): static code scan for logging
 - documentation(securityhub): security hub findings description with prescriptive guidance
 - feat(config): added deploymentTargets for awsConfig since configuration recorder cannot be turned off selectively
+- feat(guardduty): add deploymentTargets and autoEnableOrgMembers settings for guardduty
 
 ### Changed